requirements.txt
tk
Pillow
matplotlib
